Moved bootstrap/module/* to bootstrap/*
