+++ src/lib/image/bootstrap/bootstrap.c
@@ -38,8 +38,6 @@ int singularity_bootstrap_init(int argc, char ** argv) {
     ABORT(255);
   }
 
-  //Check for $SINGULARITY_ROOTFS && $SINGULARITY_libexecdir definitions
-
   //image-mount has finished, we are now inside a fork of image-mount running image-bootstrap binary instead of bootstrap.sh
 
   //mktemp -d /tmp/singularity-bootstrap.XXXXXXX ?? Unsure where this was used but was in the bootstrap shell scripts
@@ -75,7 +73,6 @@ int singularity_bootstrap_init(int argc, char ** argv) {
   return(0);
 }
 
-
 void singularity_bootstrap_script_run(char *section_name) {
   char ** fork_args;
   char *script;
@@ -102,5 +99,38 @@ void singularity_bootstrap_script_run(char *section_name) {
     free(fork_args[1]);
     free(fork_args[2]);
     free(fork_args);
+  }
+}
+
+int singularity_bootstrap_module_init() {
+  singularity_bootdef_rewind();
+
+  if ( ( module_name = singularity_bootdef_get_value("Bootstrap") ) == NULL ) {
+    singularity_message(ERROR, "Bootstrap definition file does not contain a Bootstrap: line");
+    ABORT(255);
+
+  } else {
+    singularity_message(INFO, "Running bootstrap module %s\n", module_name);
+
+    if ( strcmp(module_name, "docker") ) { //Docker
+      return( singularity_bootstrap_docker_init() );
+
+    } else if ( strcmp(module_name, "yum") ) { //Yum
+      return( singularity_bootstrap_yum_init() );
+
+    } else if ( strcmp(module_name, "debootstrap") ) { //Debootstrap
+      return( singularity_bootstrap_debootstrap_init() );
+
+    } else if ( strcmp(module_name, "arch") ) { //Arch
+      return( singularity_bootstrap_arch_init() );
+
+    } else if ( strcmp(module_name, "busybox") ) { //Busybox
+      return( singularity_bootstrap_busybox_init() );
+
+    } else {
+      singularity_message(ERROR, "Could not parse bootstrap module of type: %s", module_name);
+      ABORT(255);
+    }
+
   }
 }
