+++ src/lib/image/bootstrap/docker/docker.c
