+++ src/lib/image/bootstrap/yum/yum.c
